creating controllers, having issue with edit.ejs form to POST
