+++ views/dinosaurs/edit.ejs
@@ -1,8 +1,8 @@
-<form action="POST" action="/dinosaurs/<%= dinoId %>/?_method=PUT">
-    <label for="">Name</label>
+<form action="POST" action="/<%= dinoId %>/?_method=PUT">
+    <label for="name">Name</label>
     <input type="text" value="<%= dino.name %>" name="name" />
 
-    <label for="">Type</label>
+    <label for="type">Type</label>
     <input type="text" value="<%= dino.type %>" name="type" />
 
     <input type="submit" />
