+++ views/layout.ejs
@@ -11,10 +11,7 @@
     <nav>
         <ul>
             <li>
-                <a href="/">Dino Home</a>
-            </li>
-            <li>
-                <a href="/dinosaurs">Dinos</a>
+                <a href="/dinosaurs/index">Dino Home</a>
             </li>
             <li>
                 <a href="/dinosaurs/new">New Dino</a>
